Create ticksToRPMTest and RPMToTicksTest to test conversions

+++ src/main/java/frc/robot/Constants.java
@@ -35,7 +35,7 @@ public class Constants {
         public static final double TARGET_RPM = 500;
         public static final double PERCENT_THRESHOLD = 10;
         public static final double RAMP_RATE = 2;
-        public static final int GEAR_RATIO = 1;
+        public static final double GEAR_RATIO = 1;
     }
 
     static { // Runs alongside main
